Refactoring for MVC
