+++ controllers/tourController.js
@@ -1,17 +1,4 @@
-const fs = require('fs')
-
-const tours = JSON.parse(fs.readFileSync(`${__dirname}/../dev-data/data/tours-simple.json`))
-
-exports.checkID = (req, res, next, val) => {
-    console.log(`Tour id is: ${val}`)
-    if (req.params.id * 1 > tours.length) {
-        return res.status(404).json({
-            status: 'fail',
-            message: 'Invalid ID'
-        })
-    }
-    next()
-}
+const Tour = require('./../models/tourModel')
 
 exports.checkBody = (req, res, next) => {
     if (!req.body.name || !req.body.price) {
@@ -28,10 +15,10 @@ exports.getAllTours = (req, res) => {
     res.status(200).json({
         status: 'success',
         requestedAt: req.requestTime,
-        results: tours.length,
-        data: {
-            tours
-        }
+        // results: tours.length,
+        // data: {
+        //     tours
+        // }
     })
 }
 
@@ -39,30 +26,22 @@ exports.getTour = (req, res) => {
     console.log(req.params)
     const id = req.params.id * 1
 
-    const tour = tours.find(el => el.id === id)
+    // const tour = tours.find(el => el.id === id)
 
-    res.status(200).json({
-        status: 'success',
-        data: {
-            tour
-        }
-    })
+    // res.status(200).json({
+    //     status: 'success',
+    //     data: {
+    //         tour
+    //     }
+    // })
 }
 
 exports.createTour = (req, res) => {
-    // console.log(req.body)
-    const newId = tours[tours.length - 1].id + 1
-    const newTour = Object.assign({ id: newId }, req.body)
-
-    tours.push(newTour)
-
-    fs.writeFile(`${__dirname}/dev-data/data/tours-simple.json`, JSON.stringify(tours), err => {
-        res.status(201).json({
-            status: 'success',
-            data: {
-                tour: newTour
-            }
-        })
+    res.status(201).json({
+        status: 'success',
+        // data: {
+        //     tour: newTour
+        // }
     })
 }
